+++ Cargo.toml
@@ -1,8 +1,8 @@
 [workspace]
 members = [
-    "rushstr",         # binary crate
-    "rushstr-core",    # core logic as a library
-    "rushstr-tui",     # optional TUI interface logic
+    "rushstr", # binary crate
+    "rushstr-core", # core logic as a library
+    "rushstr-tui", # optional TUI interface logic
 ]
 
 resolver = "2"
@@ -11,3 +11,4 @@ resolver = "2"
 anyhow = "1"
 thiserror = "2"
 ratatui = { version = "0.29.0", features = ["all-widgets"] }
+uuid = { version = "1.16.0", features = ["v4"] }
