+++ rushstr-core/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2024"
 [dependencies]
 fuzzy-matcher = "0.3"
 regex = "1"
+uuid.workspace = true
