+++ rushstr-core/src/crawler/history_crawler.rs
@@ -1,0 +1,7 @@
+pub struct HistoryCrawler {}
+
+impl HistoryCrawler {
+    pub fn new() -> Self {
+        Self {}
+    }
+}
