+++ rushstr-core/src/crawler/scanner.rs
@@ -1,0 +1,5 @@
+use crate::HItem;
+
+pub trait HScanner {
+    fn load(&self) -> Vec<HItem>;
+}
