+++ rushstr-core/src/lib.rs
@@ -1,4 +1,5 @@
 mod config;
+mod crawler;
 mod stores;
 mod types;
 mod utils;
@@ -8,11 +9,17 @@ pub use config::search_options::SearchOptions;
 #[doc(inline)]
 pub use config::search_options::SearchType;
 #[doc(inline)]
+pub use crawler::history_crawler::HistoryCrawler;
+#[doc(inline)]
+pub use crawler::mem_crawler::MemCrawler;
+#[doc(inline)]
+pub use crawler::scanner::HScanner;
+#[doc(inline)]
 pub use stores::store_trait::StoreTrait;
 #[doc(inline)]
 pub use stores::vector_store::VectorStore;
 #[doc(inline)]
-pub use types::common_types::{HIndex, HLines, Store};
+pub use types::common_types::{HIndex, HLines, Store,Scanner};
 #[doc(inline)]
 pub use types::h_item::HItem;
 #[doc(inline)]
