+++ rushstr-core/src/stores/mod.rs
@@ -1,2 +1,3 @@
+pub mod matchers;
 pub mod store_trait;
 pub mod vector_store;
